requirements.txt
streamlit>=1.24.0
anthropic>=0.7.0
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
plotly>=5.8.0
pytrends>=4.8.0
requests>=2.28.0
python-dotenv>=0.20.0
google-api-python-client>=2.70.0